login

+++ src/core-domain/user-service/logoutuserinfo.service.ts
@@ -1,12 +1,14 @@
 import { Injectable } from "@nestjs/common";
 import { WinstonLoggerService } from "../../infrastructure/logger/winston-logger.service";
 import { HttpClient } from "../../infrastructure/client/http.client";
-import { UserLogoutModel } from "../models/user.logout.model";
+import { LogoutStatus, UserLogoutModel } from "../models/user.logout.model";
 import { IBaseService } from "./base.service";
+import { ResponseCode } from "../../infrastructure/constants/response-code";
 
-/* This Injectable service class contains logoutUserInfo service implementation for logout */
+
+/**  This Injectable service class contains logoutUserInfo service implementation for logout */
 @Injectable()
-export default class LogoutUserInfo implements IBaseService<UserLogoutModel, UserLogoutModel>{
+export default class LogoutUserInfo implements IBaseService<UserLogoutModel, LogoutStatus>{
     constructor(private httpclient: HttpClient,
         private logger: WinstonLoggerService,) {
         this.logger.setContext(LogoutUserInfo.name);
@@ -16,49 +18,41 @@ export default class LogoutUserInfo implements IBaseService<UserLogoutModel, Use
     /** This method receives userLogoutModel through UserController
     *   fetches user details using userLogoutModel and sends it to system api to save using httpclient post
     *   then sends userId and shopId to system api to delete user using httpcient delete*/
-    async handle(userLogoutModel: UserLogoutModel): Promise<UserLogoutModel> {
+    async handle(userLogoutModel: UserLogoutModel): Promise<LogoutStatus> {
 
-        this.logger.info('in LogoutUserInfo handle  #UserLogoutModel  ${userLogoutModel}');
-        this.logger.error('in LogoutUserInfo handle error', { key: 'value' });
-        this.logger.debug('in LogoutUserInfo handle debug', { key: 'value' });
-        this.logger.warn('in LogoutUserInfo handle warn');
+        this.logger.info(`in LogoutUserInfo handle  #UserLogoutModel  ${userLogoutModel}`);
+        this.logger.info('in LogoutUserInfo handle info', { handle: userLogoutModel });
 
-    let date: Date = new Date();
-    console.log('Date Login', date)
-    userLogoutModel.logoutDate = date;
-    
+        let date: Date = new Date();
+        userLogoutModel.logoutDate = date;
 
-    console.log('*************************',userLogoutModel)
-    const found = await this.httpclient.post('getUserInfo', userLogoutModel);
- 
-    console.log('Found',found)
+        const found = await this.httpclient.post('getUserInfo', userLogoutModel);
 
-    for (let obj of await found) {
-        console.log('Inside for for loop')
-        console.log('UserModel**(((', userLogoutModel.userId,userLogoutModel.browser, userLogoutModel.machineId,userLogoutModel.shopId)
-        console.log('%%%%%%%%%BJ',obj)
-        console.log('%-----', obj.userId, obj.browser,obj.machineId, obj.shopId )
-                if (obj.userId === userLogoutModel.userId && obj.browser === userLogoutModel.browser && 
-                    obj.machineId === userLogoutModel.machineId) {
-                    console.log('inside if block logout')
+        this.logger.info('in createUserInfo handle info found', { handle: found });
 
-                    userLogoutModel.loginDate = obj.loginDate;
- 
-   
-    console.log('User&&&')
-    const res = await this.httpclient.post('logout', userLogoutModel)
-   
-    console.log('Res++++', res)
+        for (let obj of await found) {
+            this.logger.info('in createUserInfo handle info obj', { handle: obj });
+            if (obj.userId === userLogoutModel.userId && obj.browser === userLogoutModel.browser &&
+                obj.machineId === userLogoutModel.machineId) {
 
-    const delres = await this.httpclient.delete('delUserInfo/'+userLogoutModel.userId+"/"+ userLogoutModel.shopId);
-   
-    console.log('delRes++++', delres)
+                userLogoutModel.loginDate = obj.loginDate;
 
-    return res;
-}
 
-}
+                const res = await this.httpclient.post('logout', userLogoutModel)
 
-}
+                this.logger.info('in createUserInfo handle info res', { handle: res });
+
+                const delres = await this.httpclient.delete('delUserInfo/' + userLogoutModel.userId + "/" + userLogoutModel.shopId);
+
+                this.logger.info('in createUserInfo handle info delres', { handle: delres });
+
+                const logoutStatus = new LogoutStatus(200, date, '/logout', "User logged out successfully.", ResponseCode.LOGGED_OUT)
+                return logoutStatus;
+
+            }
+
+        }
+
+    }
 
 }
